Parallel processing integrated into pipeline.

+++ pmake_opacmo.sh
@@ -1,31 +1,62 @@
 #!/bin/bash
 
-cmd=$1
-shift
+# Soft limit on maximum number of processes that should be running in parallel.
+# Since make_opacmo.sh sets STATE_NER late, there is the possibility that we
+# fork more instances than given here just because the 'scheduler' is not aware
+# of a freshly spawned process.
+max_processes=4
 
+cmd=ner
 fork=0
 
-while [[ $# -gt 0 ]] ; do
+function scheduler() {
+	local running=$max_processes
+	while [[ $running -ge $max_processes ]] ; do
+		running=0
+		for forks in fork_* ; do
+			if [ -f $forks/STATE_NER ] ; then
+				let running=running+1
+			fi
+		done
+		if [[ $running -ge $max_processes ]] ; then sleep 5 ; fi
+	done
+
+	return $running
+}
+
+if [ ! -d opacmo_data ] ; then mkdir opacmo_data ; fi
+
+for journal in input/* ; do
+	if [ ! -d "$journal" ] ; then continue ; fi
+
+	scheduler
+
 	forkdir=fork_$fork
 
 	rm -rf $forkdir
 	mkdir $forkdir
 	cd $forkdir
-	mkdir input ; cd input ; ln -s ../../input/$1 ; cd ..
+	mkdir input ; cd input ; ln -s $journal ; cd ..
 	ln -s ../dictionaries
 	mkdir tmp ; cd tmp ; ln -s ../../tmp/* . ; cd ..
 	ln -s ../bioknack
 	ln -s ../opacmo
 	mkdir opacmo_data
 
-	make_opacmo.sh $cmd $1
-
-	if [ "$cmd" = 'all' ] || [ "$cmd" = 'ner' ] ; then
-		mv opacmo_data/* ../opacmo_data
-	fi
+	echo "Processing in background: $journal"
+	make_opacmo.sh $cmd $journal &
+	sleep 1
 
 	cd ..
 	let fork=fork+1
 	shift
 done
+
+scheduler
+while [[ $? -neq 0 ]] ; do
+	sleep 5
+	scheduler
+done
+
+echo "Processed background processes: $fork"
